+++ Visualization.cpp
@@ -1,6 +1,6 @@
 #include "Visualization.h"
 
-Visualization::Visualization (int columns, int rows, CRGB * leds) {
+Visualization::Visualization (uint8_t columns, uint8_t rows, CRGB * leds) {
   this->columns = columns;
   this->rows = rows;
   this->leds = leds;
@@ -13,8 +13,8 @@ void Visualization::inititalize() {
   this->nextTime = 0;
 }
 
-int Visualization::xy2Pos (int x, int y) {
-  int pos = x * this->rows;
+uint16_t Visualization::xy2Pos (uint8_t x, uint8_t y) {
+  uint16_t pos = x * this->rows;
   if (x % 2 == 0) {
     pos = pos + y;
   } else {
